modules/tensorrt_demo/face_landmark/requirements.txt
Pillow==6.0.0
torch==1.0.1
torchvision==0.2.2
numpy==1.16.3
opencv-python==3.4.1
